requirements.txt
flask
flask-sqlalchemy
pandas
fpdf
xhtml2pdf
Pillow
pyinstaller
